+++ src/serviceBundle/services/Bitly.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace peter\components\serviceBundle;
+namespace peter\components\serviceBundle\Services;
 
 class Bitly implements Service
 {
